v2.3.1
Retargeted project to VS2026.

Change lost from previous update: Adjusted include paths in several places to be relative to `$(ProjectDir)`. Made minor changes to silence various compiler warnings.

Removed old documentation-generator project.

Added "Fireteams Enabled" to MP general settings.

Added enum constants for unused Armor Ability names.

Created a custom Qt Designer plug-in for weapon and Armor Ability picker widgets. Updated UI (player traits; loadouts) to use it.

Fixed tab ordering in a handful of gametype settings pages.

+++ native/src/ReachVariantTool/game_variants/components/megalo/opcode_arg_types/format_string.cpp
@@ -110,7 +110,7 @@ namespace Megalo {
       //
       this->tokenCount.read(stream);
       if (this->tokenCount > max_token_count) {
-         printf("Tokens value claimed to have %d tokens; max is %d.\n", this->tokenCount.to_int(), max_token_count);
+         printf("Tokens value claimed to have %d tokens; max is %d.\n", this->tokenCount.to_int(), (int)max_token_count);
          return false;
       }
       for (uint8_t i = 0; i < this->tokenCount; i++)
